update set ui, added default camera config

+++ lib/src/model/backend/implementations/libgphoto2/camera_controller.dart
@@ -219,7 +219,7 @@ class libgphoto2CameraController implements ICameraController {
       print(process.stderr);
     }
 
-    file.deleteSync();
+    if (file.existsSync()) file.deleteSync();
 
     return true;
   }
